# openapi.yaml
openapi: 3.1.0
info:
  title: Scheduling App API
  version: 1.0.0
  description: |
    REST API for the Scheduling App — helps American Airlines flight attendants
    build optimized bidding schedules from their monthly bid sheet, enforcing
    rules from the 2024 AA/APFA Collective Bargaining Agreement (CBA).

    ## Authentication
    All endpoints except `/auth/register` and `/auth/login` require a valid
    Bearer token in the `Authorization` header.

    ## Error Envelope
    All error responses use a standard JSON body:
    ```json
    { "code": "string", "message": "string", "details": {} }
    ```

    ## Pagination
    List endpoints return `{ data: [...], page_state: "token|null" }`.
    Pass `page_state` as a query parameter to fetch the next page.

    ## Requirement Traceability
    Each operation includes an `x-req-ids` extension listing the requirement
    IDs from requirements.md that it fulfills.
servers:
  - url: https://api.schedulingapp.example.com/v1
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [Users]
      summary: Get the authenticated user's profile and preferences
      x-req-ids: [REQ-008]
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"

    put:
      operationId: updateCurrentUser
      tags: [Users]
      summary: Update profile fields (base city, seniority, position, languages)
      x-req-ids: [REQ-008, REQ-009, REQ-010]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/me/preferences:
    put:
      operationId: updateDefaultPreferences
      tags: [Users]
      summary: Update the user's default scheduling preferences
      x-req-ids: [REQ-009]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Preferences"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preferences"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─── Bid Periods ─────────────────────────────────────────────────────────────
...
  /bid-periods/{bidPeriodId}/filter-presets:
    parameters:
      - $ref: "#/components/parameters/BidPeriodId"

    post:
      operationId: createFilterPreset
      tags: [Filter Presets]
      summary: Save a filter configuration for reuse
      x-req-ids: [REQ-006]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilterPresetInput"
      responses:
        "201":
          description: Preset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilterPreset"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    get:
      operationId: listFilterPresets
      tags: [Filter Presets]
      summary: List saved filter presets
      x-req-ids: [REQ-006]
      responses:
        "200":
          description: List of presets
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/FilterPreset"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /bid-periods/{bidPeriodId}/filter-presets/{presetId}:
    parameters:
      - $ref: "#/components/parameters/BidPeriodId"
      - $ref: "#/components/parameters/PresetId"

    delete:
      operationId: deleteFilterPreset
      tags: [Filter Presets]
      summary: Delete a saved filter preset
      x-req-ids: [REQ-006]
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Awarded Schedule ───────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BidPeriodId:
      name: bidPeriodId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PresetId:
      name: presetId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          examples: ["VALIDATION_ERROR", "NOT_FOUND", "PARSE_FAILED"]
        message:
          type: string
          description: Human-readable error description
        details:
          type: object
          additionalProperties: true
          description: Additional context (field errors, etc.)

    # ── Auth ───────────────────────────────────────────────────────────────────
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        profile:
          $ref: "#/components/schemas/Profile"
        default_preferences:
          $ref: "#/components/schemas/Preferences"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Profile:
      type: object
      properties:
        display_name:
          type: string
        base_city:
          type: string
          description: IATA code
        seniority_number:
          type: integer
          description: "Lower = more senior"
        total_base_fas:
          type: integer
          description: "Total FAs at this base; used with seniority_number to compute percentile"
        position_min:
          type: integer
        position_max:
          type: integer
        language_qualifications:
          type: array
          items:
            type: string
          description: "e.g. ['JP', 'SP']"
        commute_from:
          type: string
          nullable: true
          description: "IATA code of city FA commutes from (e.g. 'DCA'). Used for commute impact analysis on sequences."
        years_of_service:
          type: integer
          description: "Years of service for CBA §3.A pay rate lookup (1-13+)"
        is_reserve:
          type: boolean
          description: "true = Reserve, false = Lineholder"
        is_purser_qualified:
          type: boolean
          description: "Purser qualification per CBA §14.L"
        line_option:
          type: string
          enum: [standard, high, low]
          description: "Line of Time option: standard (70-90h), high (up to 110h), low (down to 40h) per CBA §2.EE"
    ProfileInput:
      type: object
      required: [display_name, base_city, seniority_number, total_base_fas, position_min, position_max]
      properties:
        display_name:
          type: string
        base_city:
          type: string
        seniority_number:
          type: integer
        total_base_fas:
          type: integer
        position_min:
          type: integer
        position_max:
          type: integer
        language_qualifications:
          type: array
          items:
            type: string
        commute_from:
          type: string
          nullable: true
          description: "IATA code of city FA commutes from (e.g. 'DCA')"
        years_of_service:
          type: integer
        is_reserve:
          type: boolean
          default: false
        is_purser_qualified:
          type: boolean
          default: false
        line_option:
          type: string
          enum: [standard, high, low]
          default: standard
    UpdateUserRequest:
      type: object
      properties:
        profile:
          $ref: "#/components/schemas/ProfileInput"
        default_preferences:
          $ref: "#/components/schemas/Preferences"
    Preferences:
      type: object
      description: All fields are optional; omitted fields retain their current value.
      properties:
        preferred_days_off:
          type: array
          items:
            type: integer
            minimum: 1
            maximum: 31
        preferred_layover_cities:
          type: array
          items:
            type: string
        avoided_layover_cities:
          type: array
          items:
            type: string
        tpay_min_minutes:
          type: integer
        tpay_max_minutes:
          type: integer
        preferred_equipment:
          type: array
          items:
            type: string
        report_earliest_minutes:
          type: integer
          description: Minutes from midnight
        report_latest_minutes:
          type: integer
        release_earliest_minutes:
          type: integer
        release_latest_minutes:
          type: integer
        avoid_redeyes:
          type: boolean
        prefer_turns:
          type: boolean
          nullable: true
        prefer_high_ops:
          type: boolean
          nullable: true
        cluster_trips:
          type: boolean
        prefer_holidays:
          type: boolean
          nullable: true
          description: "Prefer sequences operating on CBA §3.K holidays (100% premium)"
        prefer_speaker_sequences:
          type: boolean
          nullable: true
          description: "For language-qualified FAs: prefer Speaker sequences"
        weights:
          $ref: "#/components/schemas/PreferenceWeights"
    PreferenceWeights:
      type: object
      description: Relative importance (1-10) for each preference dimension
      properties:
        days_off:
          type: integer
          minimum: 1
          maximum: 10
        tpay:
          type: integer
          minimum: 1
          maximum: 10
        layover_city:
          type: integer
          minimum: 1
          maximum: 10
        equipment:
          type: integer
          minimum: 1
          maximum: 10
        report_time:
          type: integer
          minimum: 1
          maximum: 10
        release_time:
          type: integer
          minimum: 1
          maximum: 10
        redeye:
          type: integer
          minimum: 1
          maximum: 10
        trip_length:
          type: integer
          minimum: 1
          maximum: 10
        clustering:
          type: integer
          minimum: 1
          maximum: 10

    # ── Bid Period ─────────────────────────────────────────────────────────────
...
    FilterPreset:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        filters:
          $ref: "#/components/schemas/FilterSet"
        created_at:
          type: string
          format: date-time
    FilterPresetInput:
      type: object
      required: [name, filters]
      properties:
        name:
          type: string
        filters:
          $ref: "#/components/schemas/FilterSet"
    FilterSet:
      type: object
      properties:
        categories:
          type: array
          items:
            type: string
        equipment_types:
          type: array
          items:
            type: string
        layover_cities:
          type: array
          items:
            type: string
        language:
          type: string
          nullable: true
        duty_days_min:
          type: integer
          nullable: true
        duty_days_max:
          type: integer
          nullable: true
        tpay_min_minutes:
          type: integer
          nullable: true
        tpay_max_minutes:
          type: integer
          nullable: true
        tafb_min_minutes:
          type: integer
          nullable: true
        tafb_max_minutes:
          type: integer
          nullable: true
        block_min_minutes:
          type: integer
          nullable: true
        block_max_minutes:
          type: integer
          nullable: true
        operating_dates:
          type: array
          items:
            type: integer
        position_min:
          type: integer
          nullable: true
        position_max:
          type: integer
          nullable: true
        include_deadheads:
          type: boolean
          nullable: true
        is_turn:
          type: boolean
          nullable: true
        is_odan:
          type: boolean
          nullable: true
        is_ipd:
          type: boolean
          nullable: true
        has_holiday:
          type: boolean
          nullable: true
        is_speaker_sequence:
          type: boolean
          nullable: true
        international_duty_type:
          type: string
          nullable: true
          enum: [non_long_range, mid_range, long_range, extended_long_range]
        report_earliest:
          type: integer
          nullable: true
        report_latest:
          type: integer
          nullable: true
        release_earliest:
          type: integer
          nullable: true
        release_latest:
          type: integer
          nullable: true

    # ── Awarded Schedule ───────────────────────────────────────────────────────
...